requirements.txt.txt
streamlit
PyMuPDF
google-generativeai
sentence-transformers
faiss-cpu
numpy
